remove unnecessary mutabledata from instrument api
thread-safety guarantees were already violated with the introduction of handles (as opposed to using ids with the mutabledata maps internally before). so having this extra lock mechanism does not help the client in any way as they must implement a thread-safe way to destroy the created handles during processing regardless.